go.mod: Upgrade to a version of cty with the new "refinements" feature
This allows constraining the possible range of unknown values so that we
can opportunistically return known results even with unknown operands
sometimes, and so that we can detect certain invalid unknown operands
during planning rather than always being delayed until the apply step.

We're not yet making any specific use of it but even just upgrading cty
and HCL gives us the benefit of their own automatic refinement behaviors,
one of which we can see in the changes to the tests for the "length"
function, which know returns a refined unknown number in most cases.