core: Graph.ResourceGraph
This helper method derives from an execution graph a typically-smaller
graph which describes only the relationships between resources. This could
then be used for analyses that only care about the relationships between
resources and don't need to consider all of the other objects that might
indirectly glue the resources together.